Move upsert methods to its own folder + refactor generate_logs
